backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cors==1.0.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
filelock==3.7.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
zope.event==4.5.0
zope.interface==5.4.0
